Read opcodes from files
Current generation flow assumes that opcodes are provided
in a `cat ./opcodes | ./parse_opcodes -c' manner. However,
Windows CMD has no `cat' command, and it uses `type' instead,
so implementation of cross-platform script is complicated.

In this patch, we allow parsing opcode files directly by Python,
if their names are provided as a command line arguments, not depending
on the host shell. If no arguments are passed, script behaves as usual,
reading opcodes from the stdin.