Improved the latex table generation.
Now automatically generates the list of instruction types based
on the encodings used in the list of instructions to be printed,
instead of requiring the headers to be passed alongside the data.
Also inserts new pages whenever the table grows to such a length
that entries would spill over to the next one.