Reapply pre-commit autoformat
This little bit of code from #268 was not formatted in #292 because it was not merged yet. Basically it was a semantic conflict between the PRs, but not a text conflict, so Git can't detect it.

This can be prevented using merge queues, but it probably isn't necessary on a repo of this size.