Stop including pseudo instructions in output for Go
Go really only needs the instruction encodings for actual instructions.
Additional pseudo-encodings have since been added to the riscv-opcodes
generation, which also include instructions that are aliased to
themselves (e.g. AJALPSEUDO/AJALRPSEUDO). Instead of dealing with these
complications, stop including pseudo instructions in the output for Go
and we'll synthesise pseudo instructions that we need.

Add -pseudo which can be used to enable the generation of
pseudo-instructions, restoring the previous behaviour.