Simplify go's LUT
I am working on C support and it started being very awkward
to have to add things like funct4 and funct6 when the bits overlap
and the some of the fields are used differently or not used at all
by different encodings.

By passing the match the go assembler can do a simple bitwise or.
This was also changed to emit go fmt-ed code from the start as go fmt
now refuse to run due to the internal import.

Here is the matching CL on go's side https://go-review.googlesource.com/c/go/+/637940